# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type Locale = typeof Locale[keyof typeof Locale];


export const Locale = {
  en: 'en',
  ar: 'ar',
} as const;

export type LeadStatus = typeof LeadStatus[keyof typeof LeadStatus];


export const LeadStatus = {
  new: 'new',
  contacted: 'contacted',
  closed: 'closed',
} as const;

export interface CreateLeadRequest {
  /** @minLength 1 */
  name: string;
  email: string;
  phone?: string;
  company?: string;
  service?: string;
  /** @minLength 1 */
  message: string;
  locale?: Locale;
  website?: string;
}

export interface Lead {
  id: string;
  name: string;
  email: string;
  phone?: string | null;
  company?: string | null;
  service?: string | null;
  message: string;
  sourcePage?: string | null;
  locale: Locale;
  status: LeadStatus;
  createdAt: string;
}

export interface LoginRequest {
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface AdminUser {
  id: string;
  email: string;
  name: string;
}

export interface ArticleSummary {
  id: string;
  slug: string;
  category: string;
  coverImage?: string | null;
  featured: boolean;
  publishedAt?: string | null;
  title: string;
  excerpt: string;
}

export type ArticleDetailStatus = typeof ArticleDetailStatus[keyof typeof ArticleDetailStatus];


export const ArticleDetailStatus = {
  draft: 'draft',
  published: 'published',
} as const;

export type ArticleDetail = ArticleSummary & {
  status: ArticleDetailStatus;
  body: string;
};

export type AdminArticleStatus = typeof AdminArticleStatus[keyof typeof AdminArticleStatus];


export const AdminArticleStatus = {
  draft: 'draft',
  published: 'published',
} as const;

export type AdminArticleTranslationsItem = {
  locale: Locale;
  title: string;
  excerpt: string;
  body: string;
};

export interface AdminArticle {
  id: string;
  slug: string;
  category: string;
  coverImage?: string | null;
  featured: boolean;
  status: AdminArticleStatus;
  publishedAt?: string | null;
  translations: AdminArticleTranslationsItem[];
}

export type UpsertArticleRequestStatus = typeof UpsertArticleRequestStatus[keyof typeof UpsertArticleRequestStatus];


export const UpsertArticleRequestStatus = {
  draft: 'draft',
  published: 'published',
} as const;

export type UpsertArticleRequestTranslationsItem = {
  locale: Locale;
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  excerpt: string;
  /** @minLength 1 */
  body: string;
};

export interface UpsertArticleRequest {
  /** @minLength 1 */
  slug: string;
  /** @minLength 1 */
  category: string;
  coverImage?: string;
  featured?: boolean;
  status: UpsertArticleRequestStatus;
  translations: UpsertArticleRequestTranslationsItem[];
}

export type ListArticlesParams = {
locale: Locale;
category?: string;
};

export type GetArticleBySlugParams = {
locale: Locale;
};

export type AdminListLeadsParams = {
status?: LeadStatus;
};

export type AdminUpdateLeadStatusBody = {
  status: LeadStatus;
};