# lib/api-zod/src/generated/types/locale.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type Locale = typeof Locale[keyof typeof Locale];


export const Locale = {
  en: 'en',
  ar: 'ar',
} as const;